requirements.txt
# Web Framework
flask==3.0.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
werkzeug==3.0.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Remote Connections
paramiko==3.4.0
pywinrm==0.4.3

# Utilities
python-dotenv==1.0.0
bcrypt==4.1.1
apscheduler==3.10.4
requests==2.31.0

# Production Server
gunicorn==21.2.0
